+++ NEWS.md
@@ -6,6 +6,8 @@ in memory.
 rendering.
 - Custom renderers (`as_renderer`) are now more robust.
 - Add `limit` field to protect against large uploads.
+- Fix issue with setting custom websocket handler [#62](https://github.com/devOpifex/ambiorix/issues/62).
+- Add `engine` method on router to set custom renderers (`use` deprecated for custom renderers).
 
 # ambiorix 2.1.0